requirements.txt
streamlit>=1.28.0
mlflow>=2.8.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
evidently>=0.3.0
pytest>=7.4.0
xgboost>=2.0.0
joblib>=1.3.0
schedule>=1.2.0
plotly>=5.17.0
altair>=5.1.0
python-dotenv>=1.0.0
psutil>=5.9.0
geopy>=2.4.0
